flickflock-backend/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
requests==2.*
diskcache==5.*
google-cloud-firestore==2.7.*